Split tests in cuda
